# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "api para ver el tema del coso delcoso sobre el cosito(es de comprar cosas)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "LA api"
        }
    ],
    "paths": {
...
        "/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "aqui esta el stuff",
                "operationId": "3a39382d8e98e52fb91a8b609e5fa5cf",
                "responses": {
                    "200": {
                        "description": "aqui eta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "colocaciona otro item",
                "operationId": "641a6c1bf6b879ce0fc741a5a063097b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sku",
                                    "nombre",
                                    "descr_corta",
                                    "descr_larga",
                                    "precio_neto",
                                    "stock_actual",
                                    "stock_minimo",
                                    "stock_bajo",
                                    "stock_alto"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "PROD-001"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Mouse inalámbrico"
                                    },
                                    "descr_corta": {
                                        "type": "string",
                                        "example": "Mouse ergonómico"
                                    },
                                    "descr_larga": {
                                        "type": "string",
                                        "example": "Mouse inalámbrico asi terrible epico con colores apra mas aerodinamismo"
                                    },
                                    "precio_neto": {
                                        "type": "integer",
                                        "example": 10000
                                    },
                                    "stock_actual": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "stock_minimo": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "stock_bajo": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "stock_alto": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ya esta"
                    },
                    "422": {
                        "description": "no, no esta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/productos/{producto}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "dame el id porfa",
                "operationId": "063b2dd867f7761addcdd28de42e8a8d",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "efectivamente, aqui esta"
                    },
                    "404": {
                        "description": "este no eta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "ya se actualizo(la verda profe me quede sin ideas de nuevo)",
                "operationId": "a64b117e8e812eb1a7ee46d4217dd62d",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "se actualizo, bien ahi"
                    },
                    "404": {
                        "description": "no esta"
                    },
                    "422": {
                        "description": "Error en algo"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Elimina un producto por su ID.",
                "operationId": "c5cf8998ae00a090239c668d97ebc0b2",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado con éxito"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "el token de segiruda",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}